labs/module-02/demos/demo-06-llm-stream-endpoint/pyproject.toml
[project]
name = "llm-stram-endpoint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.2",
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
